KTOR-1767 update structure
