--- redirect-with-exception/build.gradle
@@ -1,36 +1,0 @@
-buildscript {
-    repositories {
-        mavenCentral()
-    }
-    dependencies {
-        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:1.7.20"
-    }
-}
-
-apply plugin: 'kotlin'
-apply plugin: 'application'
-
-mainClassName = "io.ktor.samples.redirectwithexception.RedirectWithExceptionApplicationKt"
-
-sourceSets {
-    main.kotlin.srcDirs = ['src']
-    main.resources.srcDirs = ['resources']
-    test.kotlin.srcDirs = ['test']
-    test.resources.srcDirs = ['testresources']
-}
-
-repositories {
-    mavenCentral()
-}
-
-dependencies {
-    implementation 'org.jetbrains.kotlin:kotlin-stdlib-jdk8:1.7.20'
-    implementation "io.ktor:ktor-server-html-builder:2.2.2"
-    implementation "io.ktor:ktor-server-default-headers:2.2.2"
-    implementation "io.ktor:ktor-server-status-pages:2.2.2"
-    implementation "io.ktor:ktor-server-call-logging:2.2.2"
-    implementation 'ch.qos.logback:logback-classic:1.2.11'
-    implementation 'io.ktor:ktor-server-netty-jvm:2.2.2'
-    implementation 'io.ktor:ktor-server-sessions-jvm:2.2.2'
-    implementation 'io.ktor:ktor-client-cio-jvm:2.2.2'
-}
