+++ redirect-with-exception/gradle.properties
