--- reverse-proxy-ws/build.gradle
@@ -1,41 +1,0 @@
-buildscript {
-    ext.kotlin_version = '1.7.20'
-    ext.ktor_version = '2.2.2'
-
-    repositories {
-        mavenCentral()
-        gradlePluginPortal()
-    }
-    dependencies {
-        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:1.7.20"
-    }
-}
-
-apply plugin: 'kotlin'
-apply plugin: 'application'
-
-mainClassName = "io.ktor.samples.reverseproxyws.ReverseProxyWsApplicationKt"
-
-sourceSets {
-    main.kotlin.srcDirs = ['src']
-    main.resources.srcDirs = ['resources']
-    test.kotlin.srcDirs = ['test']
-    test.resources.srcDirs = ['testresources']
-}
-
-repositories {
-    mavenCentral()
-}
-
-dependencies {
-    implementation 'org.jetbrains.kotlin:kotlin-stdlib-jdk8:1.7.20'
-    implementation "io.ktor:ktor-server-html-builder:2.2.2"
-    implementation "io.ktor:ktor-server-websockets:2.2.2"
-    implementation 'ch.qos.logback:logback-classic:1.2.11'
-    implementation 'io.ktor:ktor-server-netty-jvm:2.2.2'
-    implementation 'io.ktor:ktor-server-websockets-jvm:2.2.2'
-    implementation 'io.ktor:ktor-client-cio-jvm:2.2.2'
-    implementation 'io.ktor:ktor-client-cio-jvm:2.2.2'
-    implementation 'io.ktor:ktor-client-websockets-jvm:2.2.2'
-}
-
