--- rx/build.gradle
@@ -1,34 +1,0 @@
-buildscript {
-    repositories {
-        mavenCentral()
-    }
-    dependencies {
-        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:1.7.20"
-    }
-}
-
-apply plugin: 'kotlin'
-apply plugin: 'application'
-
-mainClassName = "io.ktor.samples.rx.RxApplicationKt"
-
-sourceSets {
-    main.kotlin.srcDirs = ['src']
-    test.kotlin.srcDirs = ['test']
-    main.resources.srcDirs = ['resources']
-}
-
-repositories {
-    mavenCentral()
-}
-
-dependencies {
-    implementation 'org.jetbrains.kotlin:kotlin-stdlib-jdk8:1.7.20'
-    implementation 'ch.qos.logback:logback-classic:1.2.11'
-
-    implementation "org.jetbrains.kotlinx:kotlinx-coroutines-rx2:1.6.4"
-    implementation 'io.ktor:ktor-server-netty-jvm:2.2.2'
-    testImplementation 'io.ktor:ktor-server-tests-jvm:2.2.2'
-
-}
-
