+++ rx/gradle.properties
