--- sse/build.gradle
@@ -1,29 +1,0 @@
-buildscript {
-    repositories {
-        mavenCentral()
-    }
-    dependencies {
-        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:1.7.20"
-    }
-}
-
-apply plugin: 'kotlin'
-apply plugin: 'application'
-
-mainClassName = "io.ktor.samples.sse.SseApplicationKt"
-
-sourceSets {
-    main.kotlin.srcDirs = ['src']
-    main.resources.srcDirs = ['resources']
-}
-
-repositories {
-    mavenCentral()
-}
-
-dependencies {
-    implementation 'org.jetbrains.kotlin:kotlin-stdlib-jdk8:1.7.20'
-    implementation 'ch.qos.logback:logback-classic:1.2.11'
-    implementation 'io.ktor:ktor-server-netty-jvm:2.2.2'
-}
-
