--- structured-logging/build.gradle
@@ -1,38 +1,0 @@
-buildscript {
-    repositories {
-        mavenCentral()
-    }
-    dependencies {
-        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:1.7.20"
-    }
-}
-
-apply plugin: 'kotlin'
-apply plugin: 'application'
-
-mainClassName = "io.ktor.samples.structuredlogging.ApplicationKt"
-
-sourceSets {
-    main.kotlin.srcDirs = ['src']
-    test.kotlin.srcDirs = ['test']
-    main.resources.srcDirs = ['resources']
-}
-
-repositories {
-    mavenCentral()
-    maven {
-        url "https://maven.pkg.jetbrains.space/kotlin/p/kotlin/kotlin-js-wrappers/"
-    }
-}
-
-dependencies {
-    implementation 'org.jetbrains.kotlin:kotlin-stdlib-jdk8:1.7.20'
-    implementation "io.ktor:ktor-server-html-builder:2.2.2"
-    implementation 'ch.qos.logback:logback-classic:1.2.11'
-    implementation 'net.logstash.logback:logstash-logback-encoder:7.1.1'
-
-    implementation 'org.jetbrains:kotlin-css-jvm:1.0.0-pre.156-kotlin-1.5.0'
-    implementation 'io.ktor:ktor-server-netty-jvm:2.2.2'
-    testImplementation 'io.ktor:ktor-server-tests-jvm:2.2.2'
-}
-
