+++ structured-logging/gradle.properties
