+++ version-diff/README.md
@@ -7,5 +7,5 @@ A diff tool for maven artifacts written using the [Ktor](https://ktor.io) client
 Execute this command in a sample directory:
 
 ```bash
-./gradlew run --args="1.6.8 2.1.2 <Optional repo url, default https://repo.maven.apache.org/maven2/io/ktor>"
+./gradlew run --args="1.6.8 2.2.2 <Optional repo url, default https://repo.maven.apache.org/maven2/io/ktor>"
 ```
