+++ youkube/src/main/resources/blog.css
